+++ sdk/formrecognizer/azure-ai-formrecognizer/src/main/java/com/azure/ai/formrecognizer/models/CustomFormModelInfo.java
@@ -24,35 +24,35 @@ public final class CustomFormModelInfo {
     private final CustomFormModelStatus status;
 
     /*
-     * Date and time (UTC) when the model was created.
+     * Date and time (UTC) when the training model was requested.
      */
-    private final OffsetDateTime createdOn;
+    private final OffsetDateTime requestedOn;
 
     /*
-     * Date and time (UTC) when the status was last updated.
+     * Date and time (UTC) when the model training was completed.
      */
-    private final OffsetDateTime lastUpdatedOn;
+    private final OffsetDateTime completedOn;
 
     /**
-     * Constructs a {@link CustomFormModelInfo} box object.
+     * Constructs a {@link CustomFormModelInfo} object.
      *
      * @param modelId The model identifier.
      * @param status The status of the model.
-     * @param createdOn The date and time (UTC) when the model was created.
-     * @param lastUpdatedOn The date and time (UTC) when the status was last updated.
+     * @param requestedOn Date and time (UTC) when the training model was requested.
+     * @param completedOn Date and time (UTC) when the model training was completed.
      */
-    public CustomFormModelInfo(final String modelId, final CustomFormModelStatus status, final OffsetDateTime createdOn,
-        final OffsetDateTime lastUpdatedOn) {
+    public CustomFormModelInfo(final String modelId, final CustomFormModelStatus status,
+        final OffsetDateTime requestedOn, final OffsetDateTime completedOn) {
         this.modelId = modelId;
         this.status = status;
-        this.createdOn = createdOn;
-        this.lastUpdatedOn = lastUpdatedOn;
+        this.requestedOn = requestedOn;
+        this.completedOn = completedOn;
     }
 
     /**
      * Get the model identifier.
      *
-     * @return the modelId value.
+     * @return the {@code modelId} value.
      */
     public String getModelId() {
         return this.modelId;
@@ -61,28 +61,28 @@ public final class CustomFormModelInfo {
     /**
      * Get the Status of the model.
      *
-     * @return the status value.
+     * @return the {@code status} value.
      */
     public CustomFormModelStatus getStatus() {
         return this.status;
     }
 
     /**
-     * Get the date and time (UTC) when the model was created.
+     * Get the date and time (UTC) when the training was requested.
      *
-     * @return the createdDateTime value.
+     * @return the {@code requestedOn} value.
      */
-    public OffsetDateTime getCreatedOn() {
-        return this.createdOn;
+    public OffsetDateTime getRequestedOn() {
+        return this.requestedOn;
     }
 
     /**
-     * Get the date and time (UTC) when the status was last updated.
+     * Get the date and time (UTC) when the model training was completed.
      *
-     * @return the lastUpdatedDateTime value.
+     * @return the {@code completedOn} value.
      */
-    public OffsetDateTime getLastUpdatedOn() {
-        return this.lastUpdatedOn;
+    public OffsetDateTime getCompletedOn() {
+        return this.completedOn;
     }
 
 }