+++ sdk/search/azure-search-documents/src/main/java/com/azure/search/documents/indexes/SearchIndexerClient.java
@@ -135,22 +135,41 @@ public class SearchIndexerClient {
      * @return a list of DataSources
      */
     public PagedIterable<SearchIndexerDataSourceConnection> listDataSourceConnections() {
-        return listDataSourceConnections(null, null, Context.NONE);
+        return listDataSourceConnections(null, Context.NONE);
     }
 
     /**
      * List all DataSources from an Azure Cognitive Search service.
      *
-     * @param select Selects which top-level properties of DataSource definitions to retrieve. Specified as a
-     * comma-separated list of JSON property names, or '*' for all properties. The default is all properties.
      * @param requestOptions Additional parameters for the operation. Contains the tracking ID sent with the request to
      * help with debugging.
      * @param context Additional context that is passed through the HTTP pipeline during the service call.
      * @return a response containing the list of DataSources.
      */
-    public PagedIterable<SearchIndexerDataSourceConnection> listDataSourceConnections(String select,
-        RequestOptions requestOptions, Context context) {
-        return new PagedIterable<>(asyncClient.listDataSourceConnections(select, requestOptions, context));
+    public PagedIterable<SearchIndexerDataSourceConnection> listDataSourceConnections(RequestOptions requestOptions,
+        Context context) {
+        return new PagedIterable<>(asyncClient.listDataSourceConnections(requestOptions, context));
+    }
+
+    /**
+     * List all DataSources names from an Azure Cognitive Search service.
+     *.
+     * @param requestOptions Additional parameters for the operation. Contains the tracking ID sent with the request to
+     * help with debugging.
+     * @param context Additional context that is passed through the HTTP pipeline during the service call.
+     * @return a response containing the list of DataSource names.
+     */
+    public PagedIterable<String> listDataSourceConnectionNames(RequestOptions requestOptions, Context context) {
+        return new PagedIterable<>(asyncClient.listDataSourceConnectionNames(requestOptions, context));
+    }
+
+    /**
+     * List all DataSource names from an Azure Cognitive Search service.
+     *
+     * @return a list of DataSources names
+     */
+    public PagedIterable<String> listDataSourceConnectionNames() {
+        return listDataSourceConnectionNames(null, Context.NONE);
     }
 
     /**
@@ -237,20 +256,38 @@ public class SearchIndexerClient {
      * @return all Indexers from the Search service.
      */
     public PagedIterable<SearchIndexer> listIndexers() {
-        return listIndexers(null, null, Context.NONE);
+        return listIndexers(null, Context.NONE);
     }
 
     /**
      * Lists all indexers available for an Azure Cognitive Search service.
      *
-     * @param select Selects which top-level properties of the indexers to retrieve. Specified as a comma-separated list
-     * of JSON property names, or '*' for all properties. The default is all properties.
      * @param requestOptions Additional parameters for the operation.
      * @param context additional context that is passed through the HTTP pipeline during the service call
      * @return all Indexers from the Search service.
      */
-    public PagedIterable<SearchIndexer> listIndexers(String select, RequestOptions requestOptions, Context context) {
-        return new PagedIterable<>(asyncClient.listIndexers(select, requestOptions, context));
+    public PagedIterable<SearchIndexer> listIndexers(RequestOptions requestOptions, Context context) {
+        return new PagedIterable<>(asyncClient.listIndexers(requestOptions, context));
+    }
+
+    /**
+     * Lists all indexers names for an Azure Cognitive Search service.
+     *
+     * @return all Indexer names from the Search service .
+     */
+    public PagedIterable<String> listIndexerNames() {
+        return listIndexerNames(null, Context.NONE);
+    }
+
+    /**
+     * Lists all indexers names for an Azure Cognitive Search service.
+     *
+     * @param requestOptions Additional parameters for the operation.
+     * @param context additional context that is passed through the HTTP pipeline during the service call
+     * @return all Indexer names from the Search service.
+     */
+    public PagedIterable<String> listIndexerNames(RequestOptions requestOptions, Context context) {
+        return new PagedIterable<>(asyncClient.listIndexerNames(requestOptions, context));
     }
 
     /**
@@ -427,22 +464,41 @@ public class SearchIndexerClient {
      * @return the list of skillsets.
      */
     public PagedIterable<SearchIndexerSkillset> listSkillsets() {
-        return listSkillsets(null, null, Context.NONE);
+        return listSkillsets(null, Context.NONE);
     }
 
     /**
      * Lists all skillsets available for an Azure Cognitive Search service.
      *
-     * @param select selects which top-level properties of the skillset definitions to retrieve. Specified as a
-     * comma-separated list of JSON property names, or '*' for all properties. The default is all properties
      * @param requestOptions additional parameters for the operation. Contains the tracking ID sent with the request to
      * help with debugging
      * @param context additional context that is passed through the HTTP pipeline during the service call
      * @return the list of skillsets.
      */
-    public PagedIterable<SearchIndexerSkillset> listSkillsets(String select, RequestOptions requestOptions,
+    public PagedIterable<SearchIndexerSkillset> listSkillsets(RequestOptions requestOptions,
         Context context) {
-        return new PagedIterable<>(asyncClient.listSkillsets(select, requestOptions, context));
+        return new PagedIterable<>(asyncClient.listSkillsets(requestOptions, context));
+    }
+
+    /**
+     * Lists all skillset names for an Azure Cognitive Search service.
+     *
+     * @return the list of skillset names.
+     */
+    public PagedIterable<String> listSkillsetNames() {
+        return listSkillsetNames(null, Context.NONE);
+    }
+
+    /**
+     * Lists all skillset names for an Azure Cognitive Search service.
+     *
+     * @param requestOptions additional parameters for the operation. Contains the tracking ID sent with the request to
+     * help with debugging
+     * @param context additional context that is passed through the HTTP pipeline during the service call
+     * @return the list of skillset names.
+     */
+    public PagedIterable<String> listSkillsetNames(RequestOptions requestOptions, Context context) {
+        return new PagedIterable<>(asyncClient.listSkillsetNames(requestOptions, context));
     }
 
     /**