+++ sdk/search/azure-search-documents/CHANGELOG.md
@@ -12,6 +12,8 @@
 and renamed `SearchIndexerDataSourceConnection` CRUD methods in `SearchIndexerClient` and `SearchIndexerAsyncClient`.
 - Removed `DataSourceCredential` and `AzureActiveDirectoryApplicationCredentials`, 
 and uplifted the properties to `SearchIndexerDataSourceConnection` and `SearchResourceEncryptionKey` respectively.
+- Removed `select` parameter from list service resource APIs.
+- Added list names APIs for each search service resource. (e.g. `listSearchIndexNames`, `listSearchIndexerNames`, `listDataSourceNames`, `listSkillsetNames`, `listSynonymMapNames`)
 
 ## 1.0.0-beta.3 (2020-05-05)
 - Replaced `isRetrievable` API with `isHidden`, parameter name changed from `retrievable` to `hidden`.